Response reflects kind(temporary)
Response reflects kind in JSON dictionary but will only pop it from the dictionary if kind is in it. This prevents issues when loading from persistent storage. The intention was to allow JSON objects to be sent and recreated based on their kind(i.e. response vs question vs quiz).